[REF] web_tour: reorganize the folder structure
Currently, all files related to the tour system are located in a folder
named tour_service.
To make the file structure more consistent in the web_tour addon,
we've created a structure similar to what's in /web.
