+++ addons/mail/__manifest__.py
@@ -177,11 +177,8 @@ For more specific needs, you may also assign custom-defined actions
             "mail/static/src/utils/common/format.js",
         ],
         'mail.assets_discuss_public_test_tours': [
-            'web_tour/static/src/tour_pointer/**/*',
             'web/static/lib/hoot-dom/**/*',
-            # scss not needed in tests and depends on scss variables that are not in this bundle
-            ('remove', 'web_tour/static/src/tour_pointer/**/*.scss'),
-            'web_tour/static/src/tour_service/**/*',
+            'web_tour/static/src/js/**/*',
             'web_tour/static/src/tour_utils.js',
             'web/static/tests/legacy/helpers/cleanup.js',
             'web/static/tests/legacy/helpers/utils.js',
