+++ addons/point_of_sale/__manifest__.py
@@ -204,9 +204,7 @@
         ],
         'point_of_sale.base_tests': [
             "web/static/lib/hoot-dom/**/*",
-            "web_tour/static/src/tour_pointer/**/*.xml",
-            "web_tour/static/src/tour_pointer/**/*.js",
-            "web_tour/static/src/tour_service/**/*",
+            "web_tour/static/src/js/**/*",
             'web_tour/static/src/tour_utils.js',
             "barcodes/static/tests/legacy/helpers.js",
             "web/static/tests/legacy/helpers/utils.js",
