+++ addons/web_tour/__manifest__.py
@@ -22,9 +22,7 @@ Odoo Web tours.
             'web/static/lib/hoot-dom/**/*',
         ],
         'web.assets_frontend': [
-            'web_tour/static/src/tour_utils.js',
-            'web_tour/static/src/tour_pointer/**/*',
-            'web_tour/static/src/tour_service/**/*',
+            'web_tour/static/src/js/**/*',
             'web_tour/static/src/tour_utils.js',
             'web/static/lib/hoot-dom/**/*',
         ],
