+++ addons/web_tour/static/src/js/tour_automatic/tour_helpers.js
