+++ addons/web_tour/static/src/js/tour_pointer/tour_pointer.xml
