+++ addons/web_tour/static/src/js/tour_recorder/tour_recorder_state.js
