+++ addons/web_tour/static/src/js/tour_state.js
