+++ addons/web_tour/static/src/js/tour_step.js
