+++ addons/web_tour/static/src/scss/tour_pointer.scss
